django_waveform_generator/requirements.txt
# Django Audio Waveform Generator Requirements
Django>=4.2.0
djangorestframework>=3.14.0
Pillow>=10.0.0
numpy>=1.24.0
librosa>=0.10.0
requests>=2.31.0
supabase>=2.0.0
matplotlib>=3.7.0  # Optional: for advanced waveform visualization
